I can not fix a buffer_free:
//mongoproxy_session_free(sess); // TODO: I can not fix this, we core on : 2012-09-28 23:23:13.525644 [DEBUG] <62d8c700>(common/buffer.c:39)[func:buffer_free] _buffer_free_ptr: 0x25fd5d0

do it latter

+++ common/buffer.h
@@ -16,7 +16,7 @@ typedef struct buffer_s{
 
 
 
-buffer_t * buffer_new(int len);
+buffer_t* buffer_new(size_t len) ;
 void buffer_free(buffer_t * b);
 void buffer_reset(buffer_t *b);
